+++ Managers/MessageManagers/messagehandler.h
@@ -52,7 +52,7 @@ signals:
 
     void updatingLatestMessagesFromServer(const QByteArray &latestMessages);
 
-    void sendMessage(const QString &message, const int &receiver_id, const QString &flag);
+    void sendMessage(const QString &message, const quint64 &receiver_id, const QString &flag);
     void sendMessageWithFile(const QString &message, const int &receiver_id,const QString& filePath, const QString &flag);
     void sendVoiceMessage(const int &receiver_id, const QString &flag);
